requirements.txt
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.0.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.33.2
torch==2.0.1
fastapi==0.103.1
uvicorn==0.23.2
nltk==3.8.1
scikit-learn==1.3.1